requirements.txt
Pillow>=9.0.0
pytesseract>=0.3.10
watchdog>=2.1.0
lark-oapi>=1.0.0
# smtplib # 内置，无需安装
# email # 内置，无需安装
APScheduler>=3.9.0
Jinja2>=3.1.0
PyYAML>=6.0